Ore Resourcepack for Eye Protection
This is a 64x ore eye protection resource package. The main reason for making 64x is not to make the border of the map too thick.If you have good suggestions, you can leave a message in the comments area and let me know.In addition, this resource package supports 1.13-1.19.3. You can put it in any version of 1.13.0-1.19.3 (you can ignore the annoying red border in versions other than 1.17. If you don't want to see the red border, you can also change it to the corresponding resource package version in the "pack. mcmeta" file of the resource package)
1.13 (17w48a) - 1.14.4 (19w46b) is 4 
1.15 (115-pre1) - 1.16.1 (1.16.2-pre3) is 5 
1.16.2 (116.2-c1) - 1.16.5 is 6 
1.17 (20w45a) - 1.17.1 (21w38a) is 7 
1.18 (21w39a) - 1.18.2 is 8 
1.19 (22w11a) - 1.19.2 is 9
1.19.3 Snapshot (22w42a-22w44a) is 11
1.19.3 (22W45A) and above is 12